Avoid generation of sse4 code with --enable-aesni
When configured with --enable-aesni, other extensions should not be
used, otherwise the library may not run on systems that don't support
them.  This has been verified with gcc and qemu.

When --enable-aesni, use custom CFLAGS to compile wolfcrypt/src/aes.c,
and no flags are being added to the global CFLAGS.
When compiling aes.c, gcc will only use -maes.  Other compilers flags
are being kept as they were.

If --enable-intelasm is used, the flags are kept as they were.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>